changes: commit 1 - For ExportHTML, removed unused import statements for ExportHTML, removed some code for nicer formatting of the exported table.
commit 2 - For UserController, added a helper method to print the speaker(s) of an event in single line format for the exportMyEvents method.

commit 3 - For ReadAndWrite, made some changes so that the weird characters that appeared on the top left of the exported table no longer appears for the exporttoHTML class.